refactor(comp:style): update color base
