+++ packages/site/src/components/layout/SideNav.vue
@@ -39,8 +39,8 @@
 
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue'
-import { appContextToken } from '../context'
-import { config } from '../sideNav'
+import { appContextToken } from '../../context'
+import { config } from '../../sideNav'
 
 export default defineComponent({
   setup() {
