+++ packages/site/src/components/layout/footer/Index.vue
@@ -13,7 +13,7 @@
 
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue'
-import { appContextToken } from '../../context'
+import { appContextToken } from '../../../context'
 
 import FooterCol from './Col.vue'
 import FooterItem from './Item.vue'
