+++ packages/site/src/components/layout/footer/locales.ts
