+++ packages/site/src/components/layout/header/GithubButton.vue
@@ -6,7 +6,7 @@
 
 <script lang="ts">
 import { defineComponent, inject, ref } from 'vue'
-import { appContextToken, isDevMode } from '../../context'
+import { appContextToken, isDevMode } from '../../../context'
 
 export default defineComponent({
   setup() {
