+++ packages/site/src/components/layout/header/Navigation.vue
@@ -33,7 +33,7 @@
 
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue'
-import { appContextToken } from '../../context'
+import { appContextToken } from '../../../context'
 
 export default defineComponent({
   setup() {
