+++ packages/site/src/components/layout/header/SearchBox.vue
@@ -6,7 +6,7 @@
 
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue'
-import { appContextToken } from '../../context'
+import { appContextToken } from '../../../context'
 
 export default defineComponent({
   setup() {
