+++ packages/site/src/components/views/home/Index.vue
@@ -41,7 +41,7 @@
 
 <script lang="ts">
 import { computed, defineComponent, inject } from 'vue'
-import { appContextToken } from '../context'
+import { appContextToken } from '../../../context'
 import { homeLocales } from './locales'
 
 export default defineComponent({
