+++ packages/site/src/components/views/home/locales.ts
