+++ packages/site/src/styles/layout/footer.less
@@ -3,12 +3,12 @@
   padding: @spacing-4xl @spacing-4xl @spacing-2xl;
   font-size: @font-size-md;
   line-height: @line-height-base;
-  color: @grey-d40;
+  color: @grey;
   background-color: @primary-d40;
 
   & a {
     transition: @transition-all-base;
-    color: @grey-d40;
+    color: @grey;
     text-decoration: none;
     &:hover {
       color: @grey-l20;
@@ -26,7 +26,7 @@
     margin-bottom: @spacing-2xl;
 
     h3 {
-      color: @grey-d20;
+      color: @grey-l10;
     }
 
     &-icon {
@@ -52,7 +52,7 @@
     text-align: center;
     padding-top: @spacing-lg;
     font-size: @font-size-lg;
-    border-top: @border-width-sm @border-style @grey-d40;
+    border-top: @border-width-sm @border-style @grey;
   }
 }
